gateway/tests/requirements.txt
pytest
pytest-html
requests
python-docx
openpyxl
Pillow
reportlab
